+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Custom feeds where articles are added manually.
 - Setting feed generation mode to every episode only generates audio automatically for new articles.
+- Changing episode play button to a generate button when audio hasn't been generated yet.
 
 ## V1.0.5